Always make the libdarknet.so
Since the darknet.py library depends on having the `libdarknet.so`
dynamic library, and we (globus webapp) depend on darknet.py, it
might be a good idea to actually build the `libdarknet.so`.

This line might've been commented out at some point since I
originally did not have this issue.